update tests and run script
