+++ src/main/java/org/matsim/dashboard/LeipzigDashboardProvider.java
@@ -2,7 +2,7 @@ package org.matsim.dashboard;
 
 import org.matsim.application.ApplicationUtils;
 import org.matsim.core.config.Config;
-import org.matsim.run.RunLeipzigScenario;
+import org.matsim.run.LeipzigScenario;
 import org.matsim.simwrapper.Dashboard;
 import org.matsim.simwrapper.DashboardProvider;
 import org.matsim.simwrapper.SimWrapper;
@@ -23,7 +23,7 @@ public class LeipzigDashboardProvider implements DashboardProvider {
 		TripDashboard trips = new TripDashboard("mode_share_ref.csv", "mode_share_per_dist_ref.csv", "mode_users_ref.csv");
 
 		return List.of(trips,
-			new TravelTimeComparisonDashboard(ApplicationUtils.resolve(config.getContext(), "leipzig-v" + RunLeipzigScenario.VERSION + "-routes-ref.csv.gz"))
+			new TravelTimeComparisonDashboard(ApplicationUtils.resolve(config.getContext(), "leipzig-v" + LeipzigScenario.VERSION + "-routes-ref.csv.gz"))
 		);
 	}
